requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
airsim==1.8.1
anyio==3.7.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
async-timeout==4.0.3
asynctest==0.13.0
attrs==23.1.0
autocfg==0.0.8
autograd==1.6.2
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
certifi==2023.11.17
cffi==1.15.1
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
cma==2.7.0
colorama==0.4.6
comm==0.1.4
cycler==0.11.0
debugpy==1.7.0
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.7
docker-pycreds==0.4.0
docopt==0.6.2
dominate==2.9.0
entrypoints==0.4
exceptiongroup==1.2.0
fastjsonschema==2.19.0
filelock==3.12.2
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2023.1.0
future==0.18.3
gitdb==4.0.11
GitPython==3.1.43
gluoncv==0.10.5.post0
graphviz==0.8.4
imageio==2.31.2
importlib-metadata==6.7.0
importlib-resources==5.12.0
ipython==7.34.0
ipython-genutils==0.2.0
ipywidgets==8.1.1
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.17.3
kiwisolver==1.4.5
lightning-utilities==0.10.1
llvmlite==0.39.1
MarkupSafe==2.1.3
matplotlib==3.3.3
matplotlib-inline==0.1.6
mistune==3.0.2
msgpack-python==0.5.6
msgpack-rpc-python==0.4.1
multidict==6.0.5
mxnet==1.7.0.post2
mxnet-cu101==1.5.0
mxnet-cu102==1.7.0
mxnet-cu92==1.5.0
nest-asyncio==1.5.8
numba==0.56.4
numpy==1.21.6
opencv-contrib-python==4.8.1.78
opencv-python==4.8.1.78
packaging==23.2
pandas==1.1.5
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.5.0
pipenv==2023.10.3
pipreqs==0.4.13
pkgutil_resolve_name==1.3.10
platformdirs==4.0.0
portalocker==2.7.0
prometheus-client==0.17.1
prompt-toolkit==3.0.41
protobuf==4.24.4
psutil==5.9.6
pycparser==2.21
Pygments==2.17.2
pymoo==0.4.2.2
pynndescent==0.5.13
pyparsing==3.1.1
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
pyrsistent==0.19.3
python-dateutil==2.8.2
pytorch-lightning==1.9.5
pytz==2023.3.post1
PyWavelets==1.3.0
PyYAML==6.0.1
pyzmq==24.0.1
qtconsole==5.4.4
QtPy==2.4.1
scikit-learn==0.24.2
scipy==1.7.3
seaborn==0.12.2
Send2Trash==1.8.2
sentry-sdk==1.9.0
setproctitle==1.3.3
sip==6.8.3
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.4.1
threadpoolctl==3.1.0
tifffile==2021.11.2
tinycss2==1.2.1
tomli==2.0.1
torch==1.10.0+cu113
torchaudio==0.10.0+cu113
torchmetrics==0.11.4
torchvision==0.11.1+cu113
tornado==4.5.3
tqdm==4.66.1
traitlets==5.9.0
typing_extensions==4.7.1
umap-learn==0.5.3
urllib3==1.22
virtualenv==20.24.7
wandb==0.17.3
wcwidth==0.2.12
webencodings==0.5.1
websocket-client==1.6.1
widgetsnbextension==4.0.9
yacs==0.1.8
yarg==0.1.9
yarl==1.9.4
zipp==3.15.0
